refactor(audit-dialog): move file row into _components/tabs folder
